# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "oudoumariko@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/inventaires": {
            "get": {
                "tags": [
                    "inventaire"
                ],
                "summary": "Get inventaires information",
                "description": "Returns data",
                "operationId": "getAllInventaire",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/inventaire": {
            "post": {
                "tags": [
                    "inventaire"
                ],
                "summary": "Create inventaire",
                "description": "Returns response",
                "operationId": "createInventaire",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Inventaire"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/inventaire/{id}": {
            "get": {
                "tags": [
                    "inventaire"
                ],
                "summary": "Get inventaire information",
                "description": "Returns data",
                "operationId": "getByIdInventaire",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "inventaire id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "inventaire"
                ],
                "summary": "Update existing inventaire",
                "description": "Returns response",
                "operationId": "updateInventaire",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "inventaire id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Inventaire"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/inventaires/{id_periode_inventaire}": {
            "get": {
                "tags": [
                    "inventaire"
                ],
                "summary": "Get inventaire information",
                "description": "Returns data",
                "operationId": "getByIdPeriodeInventaire",
                "parameters": [
                    {
                        "name": "id periode inventaire",
                        "in": "path",
                        "description": "inventaire id_periode_inventaire",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/inventaires/codification/{id_codification}/periodeinventaire/{id_periode_inventaire}": {
            "get": {
                "tags": [
                    "inventaire"
                ],
                "summary": "Get inventaire information",
                "description": "Returns data",
                "operationId": "getInventaireByCodificationAndPeriodeInventaire",
                "parameters": [
                    {
                        "name": "id codification",
                        "in": "path",
                        "description": "inventaire id_codification",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id periode inventaire",
                        "in": "path",
                        "description": "inventaire id_periode_inventaire",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Inventaire": {
                "title": "Inventaire",
                "description": "",
                "properties": {
                    "id": {
                        "title": "Identifiant",
                        "type": "integer"
                    },
                    "etat": {
                        "title": "etat",
                        "type": "string"
                    },
                    "nom_agent": {
                        "title": "Nom agent",
                        "type": "string"
                    },
                    "observations": {
                        "title": "Nom agent",
                        "type": "string"
                    },
                    "id_codification": {
                        "title": "codification ",
                        "type": "string"
                    },
                    "id_periode_inventaire": {
                        "title": "Identifiant inentaire periode",
                        "type": "string"
                    },
                    "userId": {
                        "title": "Identifiant user",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Inventaire"
                }
            },
...
        }
    },
...
}